+++ fedmsg_meta_fedora_infrastructure/compose.py
@@ -19,6 +19,8 @@
 #
 from fedmsg_meta_fedora_infrastructure import BaseProcessor
 
+from datetime import datetime
+
 
 class ComposeProcessor(BaseProcessor):
     __name__ = "compose"
@@ -35,6 +37,8 @@ class ComposeProcessor(BaseProcessor):
             # Some old messages in datanommer don't have that branch field,
             # so we have to extract it from the topic.
             branch = msg['topic'].split('.')[-3]
+            if branch == 'compose':
+                branch = msg['topic'].split('.')[-2]
 
         arch = msg['msg'].get('arch', '')
         arch = arch and ' (%s)' % arch
@@ -72,12 +76,15 @@ class ComposeProcessor(BaseProcessor):
             base = "https://dl.fedoraproject.org/pub/" + \
                 "fedora-secondary/development"
         else:
-            base = "https://dl.fedoraproject.org/pub/" + \
-                "fedora/linux/development"
-
-        # For backwards compatibility (with old messages in datanommer)
-        if 'rawhide' in msg['topic']:
-            return base + "/rawhide"
+            # For backwards compatibility (with old messages in datanommer)
+            if 'rawhide' in msg['topic']:
+                return "http://kojipkgs.fedoraproject.org/mash/rawhide-" + \
+                    datetime.fromtimestamp(msg['timestamp']).strftime('%Y%m%d')
+            elif 'epelbeta' in msg['topic']:
+                base = "https://dl.fedoraproject.org/pub/epel/beta/7"
+            else:
+                base = "https://dl.fedoraproject.org/pub/" + \
+                    "fedora/linux/development"
 
         return base + "/" + msg['msg'].get('branch', '')